refactor:循環参照の排除
