+++ server/src/adapter/serializer/user/user_serialize.ts
@@ -1,4 +1,4 @@
-import { BaseSerialize, Response } from "../"
+import { BaseSerialize, Response } from "../base_serialize"
 import { User } from "../../../domain"
 import CONSTANT from "../../../constant"
